web/backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
alembic>=1.12.1
python-jose[cryptography]>=3.3.0
passlib==1.7.4
bcrypt==3.2.0
email-validator>=2.0.0
python-multipart>=0.0.6

# For RYM URL parsing
requests>=2.28.0
beautifulsoup4>=4.11.1

# Note: Likely unused now if not interacting with Sheets
google-api-python-client>=2.100.0
google-auth>=2.23.3
pandas                             # Note: Likely unused now if not importing from CSV/Excel
numpy                              # Note: Likely unused now if not importing from CSV/Excel

# Required for packages that relied on the removed distutils module in Python 3.12+
setuptools>=65.5.1

# Pin aioredis to avoid Python 3.13 TimeoutError conflict with fastapi-admin's dependency
aioredis==1.3.1

# Note: Likely unused now if not interacting with Sheets
google-api-python-client>=2.100.0
google-auth>=2.23.3
pandas                             # Note: Likely unused now if not importing from CSV/Excel
numpy                              # Note: Likely unused now if not importing from CSV/Excel
psycopg2-binary                   # Keep for potential future DB change
email-validator>=2.1.0            # For email validation
fastapi-admin>=0.8.0              # Add fastapi-admin
# Pin aioredis to avoid Python 3.13 TimeoutError conflict with fastapi-admin's dependency
aioredis==1.3.1 